Fixed broken cookie and crumb code, presumbly due to Yahoo backend changes.
This is not working any more with httpShouldSetCookies false.
It needs to be turned on.
Yahoo is using both the cookies and the crumb together.
With this off, I can get the crumb, but requests using the crumb will be
'Unauthorized' do to invalid crumb.  If I go to the getcrumb URL at the same
time in a web browser, I see a different crumb than the one fetched here
(and it persists for awhile hitting refresh),
further indicting that the cookie and crumb are linked at the hip.
Whatever workaround code was doing manually with cookies no longer works.
I don't know enough about the old cookie handling to remove it, so it remains, and could cause other problems.

I also changed the crumb URL to
https://query2.finance.yahoo.com/v1/test/getcrumb
I explain this more in the comments.
There is also commented out code that uses the old method, but changes
the regex to deal with Yahoo's changes to how they write the crumb now.
I also include a request for 1d to reduce the amount of data that needs
to be sent back considerably. But getcrumb is the most efficient of all.

I also changed the header and made the getcrumb also use the header
code. In the past, Yahoo has giving me problems with the header,
so this was one of the first things I tried. I didn't revert because I
know this header is valid, as it is used in other libraries.