Moved struct Quote into its own file outside of OptionsChain and shared it with the new API I just implemented called latestQuotes.
latestQuotes is somewhat of a more effient replacement for recentDataBy.
This is more efficent than recentDataBy(...) for multiple reasons.
1. This API transmits less data back than recentDataBy because the latter uses the Yahoo chart API which seems to send back a lot of additional historical stuff that this library happens to throw away.
2. This API is designed for submitting multiple identifiers in one request.
Since Yahoo may limit or ban you for too much usage, it is worth considering minimizing your traffic, hence the reason I implemented this API instead of just relying on recentDataBy.

latestQuotes also seems to support options contract symbols as identifiers.

Additionally, reviewing the overall library called, I noticed the parameter use of
"symbols", which appears to be incorrect. That parameter is for listing
multiple symbols, but since all the other APIs already directly put the
symbol into the URL or use the parameter "symbol", the usage of
"symbols" seems incorrect. So I commented out those parts.

I also made some minor changes to the Options data structures and init,
to be more resilient for missing fields. I refactored the Date
conversion to return nil instead of a Date at TimeInterval 0.
And I added a few more fields for when a contractSymbol is specified
directly.